SG-40472 - Disable events filter for participants when MultiUser is ON  (#926)
SG-40472 - Disable events filter for participants when MultiUser is ON 

### Linked issues
n/a

### Summarize your change.
Commented out the part where OpenRV disable the some toolbars buttons
based on the Live Review filtering. It will be rework soon.

Before merging this PR, a PR for the Live Review plugin needs to be
merged.

### Describe the reason for the change.
We need the toolbar actions to be enabled on all time for now.

### Describe what you have tested and on which operating system.
MacOS

Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>