(Enhancement) - Support for ffmpeg 8 (including openAPV) (#947)
Support for ffmpeg-8.0, and OpenAPV support
This adds support for ffmpeg-8.0. It also updates a couple of minor API
changes to ffmpeg that are in the newer release, along with some small
changes to make OpenAPV work well.

Describe the reason for the change.
Getting OpenRV able to support OpenAPV now that its built into ffmpeg.

Describe what you have tested and on which operating system.
This has been tested on OSX and windows.

---------

Signed-off-by: Sam.Richards@taurich.org <Sam.Richards@taurich.org>