SG-40541: fix broken qmessagebox on qt 6.5+ on macOS Tahoe with new QAlertPanel (#933)
### Linked issues
Fixes SG-40541

### Summarize your change.
on macOS Tahoe, QMessageBox which is used for the mu alertPanel command.
on Tahoe, it produces a correct-looking messagebox, but it is
unresponsive to mouse clicks, keyboard TABs, button focus changes, etc.

* We tried to strongFocus the buttons: didn't work.
* We tried modifying the plist to disable liquid glass macOS look:
didn't work.
* We tried a compatibility mode setting: didn't work.
* We tried a few other things, didn't work.

In the end, we just decided to implement QAlertPanel for the purpose of
responding to the mu alertPanel command, respecting the same contract
and API of QMessageBox, but without the bugs.

The upshot is that at least the look is the same as the RV skin, which
is good.

### Describe the reason for the change.

broken Qt QMessageBox which we used to display error/info/warning
messages.

### Describe what you have tested and on which operating system.

macOS, Tahoe. This remains to be tested on Sequoia, Linux and Windows.

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

<img width="437" height="263" alt="image"
src="https://github.com/user-attachments/assets/92096da4-3e9b-4d51-93ac-801e674edf9f"
/>

<img width="608" height="269" alt="image"
src="https://github.com/user-attachments/assets/eb892364-cb88-44a4-a61b-1fc7a06092ef"
/>

---------

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>
Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>