Fixes to allow compilation on macOS
This fixes rvbootstrap failing to compile and build on macos (and perhaps other OS)

It fails building OIIO because it finds my homebrew installed ffmpeg 8 (recently supported in RV) and hits the issue that avcodec_close has been removed in favor of avcodec_free_context.  This was fixed in OIIO 2.5.19.1 https://github.com/AcademySoftwareFoundation/OpenImageIO/releases/tag/v2.5.19.1

An upgrade to OIIO subsequently requires upgrades to Imath to 3.2.2 and OpenEXR 3.4.2

As for the openexr patch file, I am currently unclear on the logic and if it is required with 3.4.  I have left a stub file in place for future use and to keep the cmake logic intact.

OIIO 2.5 is EoL and we should be moving to OIIO 3.1 however I have a feeling that could be a bigger upgrade so have left that for another commit.

Signed-off-by: Michael Oliver <mcoliver@gmail.com>