SG-40730: Fix annotations on Wacom Intuos Pro (#948)
### [SG-40730](https://jira.autodesk.com/browse/SG-40730): Fix
annotations on Wacom Intuos Pro

### Summarize your change.

Add a fallback option to check if the pointer type is a pen or an eraser
when the device type is unknown until we can update to a newer version
of Qt where the Wacom Intuos Pro will be detected properly on macOS.

### Describe the reason for the change.

The Wacom Intuos Pro tablet on macOS is not detected properly with
Qt6.5.3 and returns an Unknown device type. Note that the device is
being detected properly with Qt5. and other models of Wacom tablets are
working without any issues with Qt6. The issue is really specific to
this model of Wacom tablet and version of Qt.

Since the annotate mode is listening to stylus-pen and stylus-eraser
events when strokes are being drawn, if a pen or eraser pointer type is
detected on an unknown device type, we map them directly to the stylus
device type as a fallback.

### Describe what you have tested and on which operating system.

Drawing annotations with the Wacom Intuos Pro tablet was tested on macOS
15.6.1

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>
Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>