SG-40184: Cancel clear if plugin refuses (#961)
### Linked issues

SG-40184

### Summarize your change.

Super simple change where we return early if an event handler catches
the before-clear event and returns "cancel".

### Describe the reason for the change.

We dont usually check for confirmation before clearing the session, but
in upcoming live review changes it will pop up a confirmation dialog if
a review session is in progress.

### Describe what you have tested and on which operating system.

macOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>
Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>