SG-40180: Fix rendering anomalies of concurrent paint strokes (#957)
### SG-40180: Fix rendering anomalies of concurrent paint strokes

### Linked issues
NA

### Describe the reason for the change.
When multiple parciticpants were annotating at the same time in a Live
Review session, sometimes some rendering anomalies could be observed
where some paint strokes' points were missing until a new paint stroke
was introduced.

### Summarize your change.

**Problem:** 
In the paint commands rendering method (ImageRenderer::renderPaint()),
an FBO Paint cache is used, excluding the very last paint command
because the original code assumed that this paint stroke for example (a
PolyLine) could still be being drawn. However, this assumption was not
taking into account that multiple participants could be drawing strokes
at the same time, thus leading to occasional paint commands rendering
anomalies where a stale cached FBO would be used where it shouldn't. The
source of the problem is

**Solution:**
Now forcing to recompute the renderIDWithPartialPaint id when rendering
paint command to make sure that the correct render id (a unique
identifier associated with the node to be rendered) is used to fetch the
correct FBO cache.

**Also in this commit:**

- When deleting a paint stroke, it still remained in the RV Session
because the .splat property had not been deleted (but all the other
were:
Example: 
```
"pen:2:1:labergb_23473"
    {
        int splat = 0
    }

```
This was fixed in this commit (annotate_mode.mu).

- Each paint command was unnecessaily duplicated. This must be the
result of a bad merge in the oirginal Create branch (For reference:
https://git.autodesk.com/media-and-entertainment/shotgun-rv/blob/e012b69857c06f9da6f48b0ee948c3ce77ea2f7a/src/lib/ip/IPBaseNodes/PaintIPNode.cpp#L514)
This was fixed in this commit as well (PaintIPNode.cpp)


### Describe what you have tested and on which operating system.
Successfully tested on macOS.

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>