SG-40929: Continued fixes based on QA. (#963)
### Summarize your change.

It turns out there were a few big holes in my previous PR for SG-40182.

* in rvui.mu and elsewhere, many bind() calls for keyboard events _(or
mouse inputs)_ that dont have equivalent menu item also needed
categorizations. There was a missing bind() overload in app_utils.mu. to
do this so it was added. This is by far the largest change.

* in app_utils.mu, calling the "compositeStateFunc()" could return
garbage (yet another mu bug, which I'm not fixing here) causing the
event func to be called even if the category was disabled. It was
replaced with a more straightforward check. (the bug seemed to happen
mostly if the handler func defined in menuItem() started with a leading
"~" before its name, eg, "~togglePlay")

* in mode_manager.mu, needed to add equivalent protections to
filterLiveReviewEvents() which got removed without an equivalent in the
shuffle.

* in annotate_mode.mu, needed to make disappear the paint tools if we
are in LR presenter mode (eg: annotate_categories is disabled)

* in scrub_offset.mu, needed to guard on scrubbing if playcontrol
category is disabled.

* more filterLiveReviewEvents were removed, and replaced with the
appropriate binding (or alternative to bind funcs)

---

Things to test,

*_when in viewer mode:...._*

- fix: stepForward 1 or 100 frames is now blocked
- fix: stepBackward 1 or 100 frames is now blocked
- fix: motion scope popup menu is blocked
- fix: clicking and dragging in timeline/ motion scope is blocked
- fix: marking frames in timeline is blocked
- fix: setting point in timeline is blocked
- fix: annotate panel is closed
- fix: cannot scrub in scrub_offset mode
- fix: rotate ibage is blocked
- fix: mousewheel to scrub is disabled
- fix: dragging in main view to scrub is blocked
- fix: doing space to play is blocked.

viewmode categories:
ViewMode Categories were off in viewer mode before, but enabled now.
(toggle this category in live_review.py)

 viewmode key shortcuts:
- fix: kbd keys are blocked: M, alt-r, alt-l, shift-l, shift-r, ctrl-m,
down, p, pgup, pgdn.
- fix: kbd keys are blocked : aldkspmzvh,.<>OSorcR/

when in nudge mode...
- fix: kbd keys are blocked : left,right,up,down, control and shift
variations.
- fix: kbd keys are blocked : z, ctrl-z, alt-n,

### Describe the reason for the change.

Stephen reported a few problems reported with being able to do some
things in LR that he shouldnt be able to do when in muc presenter mode.
(participant = viewer)

### Describe what you have tested and on which operating system.

macOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

---------

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>
Signed-off-by: Cédrik Fuoco <cedrik.fuoco@autodesk.com>