Fix building OTIO wheel on windows (#972)
### Summarize your change.
This fixes building the opentimelineio wheel on Windows builds. In PR
https://github.com/AcademySoftwareFoundation/OpenRV/pull/910 we pinned
OTIO to the commit just before OTIO upgraded their pybind dependency,
since this broke the build of RV. To unblock us from upgrading the OTIO
version, this PR introduces a fix for that issue.

The problem with the pybind upgrade is that pybind changed the way it
was finding python at it was no longer finding out version of Python.
The fix is to use the CMake envars for the include path, lib path and
executable path so the pybind can find our python location.

I updated our requirements.txt file to use the latest main branch from
OTIO, as it was doing before.

### Describe the reason for the change.

So we can use the latest OTIO version

### Describe what you have tested and on which operating system.

Tested building on Windows 11.

---------

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>