SG-39445: Enable participants to broadcast annotations in Collaborative mode for OTIO-based Live Review sessions (#960)
### [SG-39445](https://jira.autodesk.com/browse/SG-39445): Enable
participants to broadcast annotations in Collaborative mode for
OTIO-based Live Review sessions

### Summarize your change.

- Add pydantic to validate payload formats in the Live Review plugin
- Add `setCurrentAnnotateModeNode` in rv.commands to manually set the
Paint node to use when manipulating stokes in annotate_mode
- Set the mode property on all strokes even if the mode is
`RenderOverMode` to prevent the code from throwing an exception when
trying to get the value. This is necessary to know if the user is
erasing or painting before sending an annotations payload.
- Replace the hardcoded values used to set the width of the stroke with
a conversion method from OTIO coordinate to the World Coordinate System
in RV.

### Describe the reason for the change.

All the changes mentioned above are needed to properly enable
participants to broadcast annotations in Collaborative mode for
OTIO-based Live Review sessions.

### Describe what you have tested and on which operating system.

Drawing strokes with the pen and the eraser tools while being a
participant of an OTIO-based Live Review session with Collaborative mode
enable. The tests were done on macOS 15.6.1

---------

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>