Fix help menu (#977)
### Fix help menu

### Summarize your change.

Remove `.items` in the menuList of the "Help" submenu. Note that the
same line without the `.items` is in the Commercial RV version of the
help menu mode file.

### Describe the reason for the change.

When launching Open RV, the following error was being displayed in the
console: `openrv_help_menu_mode.mu, line 157, char 42: Unresolved member
reference to "items" in type "MenuItem[]"`

### Describe what you have tested and on which operating system.

Launching Open RV on macOS 15.6.1 with the fix removed the error in the
console.

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>