Qt Logic cleanup based on CY year
Currently it was setting RV_VFX_PLATFORM based on finding Qt.  The logic should be to set the Qt version based on the CY requirements.  We also need to support 6.8 for CY2026 and the current logic prevented that.  Also added a supporting function to ask for the CY year when sourcing rvcmds if RV_VFX_PLATFORM has not been previously set.

Signed-off-by: Michael Oliver <mcoliver@gmail.com>