Unify OTIO versions, pin to 0.18.1 (#982)
### Pin OTIO to 0.18.1

### Summarize your change.
Currently OTIO is 0.16 on Window debug and the latest main branch for
everything else. This branch does two things:
- unify the version between all OSes
- pin the version to 0.18.1.

Windows debug builds the OTIO wheel from source since there are no
pre-built debug Windows wheels available in pypi, and we can't use the
release one since Windows doesn't allow mixing release and debug
libraries. As a result, it's version was specified outside of the
requirements.txt file we use for everything else to install from pypi.
It's version was far behind the other OSes likely because whenever we
updated it, we thought doing so in a single place would take care of
everything.

To remove the chance of this happening again, I removed the
requirements.txt file we use for installing our python build
dependencies and instead auto-generated it from a template, using the
opentimelineio version as an input variable from CMake so the version is
now shared between all builds.

Lastly, I pinned the OTIO version to 0.18.1 which was just released this
week. It's risky to use the latest main branch since one bad commit
could break OpenRV.

### Describe the reason for the change.

Build stability and consistency.

### Describe what you have tested and on which operating system.

Building on  Windows + Linux

---------

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>