SG-40369: SequenceNode Stability Improvements (#848)
### Summarize your change.

This PR fixes program crashes that can occur if you modify Sequence edl
properties. Most of these are due to lack of proper bounds checking on
the properties in various locations.

- Add bounds checks before trying to access the edl properties
- Check source indexes are within range of nodes inputs.
- Add NULL check to ImageRenderer

We also modified the node’s behavior so that it no longer throws an
exception when it enters a bad state. Instead, it now returns a
newNoImage containing the error message. We discovered that when EDL
properties are changed dynamically, the node can easily enter a
temporary, inconsistent state. This happens because property change
events may occur between the setting of individual EDL properties, and
the operations that can take place in between are not trivial to manage.
Displaying the error console in these situations proved to be both
frustrating and confusing for users.

### Describe what you have tested and on which operating system.
Linux, macOS

---------

Signed-off-by: Mark Reid <markreid@netflixanimation.com>
Signed-off-by: pbergeron-adsk <patrick.bergeron@autodesk.com>
Co-authored-by: pbergeron-adsk <patrick.bergeron@autodesk.com>