SG-41066 - Fix qalert panel textbox and QPushButton's focused state highlighting (#981)
### Linked issues

SG-41066

### Summarize your change.

The alert panel still had issues with sizing the dialogbox. In reality
the problem was simple: just set the size for the label and the layout
engine will do the rest.

Also real strange, but the focused state had lost its styling for some
reason? the focus state was not handled correctly in the qss. I added
changes to fix the problem.

Finally, the QPushButton had its text margin set in QPushButton, but for
some reason got set to "narrower" for the "default" or disabled state.
This made no sense, so I removed these two lines in the qss

### Describe the reason for the change.

UI glitches

### Describe what you have tested and on which operating system.

macOS

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>
Signed-off-by: pbergeron-adsk <patrick.bergeron@autodesk.com>