Return to building OTIO from source (#991)
### Fix occasional segault on Windows release startup

### Summarize your change.

In commit
https://github.com/AcademySoftwareFoundation/OpenRV/commit/93819da3a2af14c63fcb00c1b61fc3d3e4394c2f
to pin the OTIO version, I change the requirements.txt file used to
install OTIO to point to a version number instead of a git commit like
it was doing before.

This introduced an unintended side effect. When you point to a
particular git commit the wheel will always be built locally before
installing, which makes sense when you think about it because there are
no pre-built wheels hosted in git. Conversely, when you use a version
number, pip will check pypi for a pre-built wheel before building and
install that if it exists. The only pre-built wheel that has a match is
Windows release -- there are no debug versions, we don't use the same
flavour of Linux as OTIO pre-built wheels nor the same verison of mac
OS. This causes a problem because we are building our own version of
python and unless the compilers, shared lib versions, etc match exactly
there is a potential for ABI in compatibilities and resulting crashes.

To fix it, we will simply go back to using a git commit so we always
build our own wheels. However, we will use the version to to match the
tag in git so we are still pinned which was the point of the previous
commit.

Lastly, there was a portion of custom code for Windows debug version to
build OTIO for the exact problem we're trying to solve here. However,
since we started pointing at a git commit this all became useless since
once requirements.txt is already doing exactly the same thing. Thus, I
removed all the code and we will only have a single code path to build
OTIO.

### Describe the reason for the change.

Fix an occasional startup crash on Windows (depending on your compiler
and shared lib versions).

### Describe what you have tested and on which operating system.

Windows 11.  Built and launched successfully.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>