SG-41117: cmake use cy files for dependency versions and add experimental support CY2025 and CY2026 (#985)
### Summarize your change.
Moved hardcoded versions for various dependendencies out of individual
cmake files and into a centralized file based on CY year. There were
also changes made to allow for years beyond 2024 rather than having to
hardcode values for each year in various build files throughout the
repo.

### Describe the reason for the change.
The current codebase has checks for 2023 or 2024 littered across many
files and it makes it difficult to understand what needs to be updated
for each CY year and where you do it. The code was also difficult to
follow with macros defining vars for various years.

With this setup you can create a new CYYEAR.cmake file, copy the old
one, update version numbers and even diff between them. It makes it much
easier to understand where to go to update packages. Additionally it
allows for things like CY2099.cmake which can have the most current
releases of packages (perhaps ahead of a CY official designation) for
people who like to run the latest dependencies with performance
improvements and bug fixes.

### Describe what you have tested and on which operating system.
Builds 2023, 2024, and 2025 have been verified on macOS.  

### Add a list of changes, and note any that might need special
attention during the review.
- Recently there was an OTIO commit that may conflict with this PR.
- CY2026 is still experimental. QtWebEngine no longer is packaged with
6.8.3 so we have to go get it.
- patch files still need to be generated for new versions of openexr and
python (though I would love to discuss why the patch file for openexr
exists. Covering up bad data in a file with black seems like it could
result in confusion for users. The file should be invalid if part of it
is invalid.)
- I also added a helper function to rvcmds to select the CY year you
want to build.
- Finding Qt logic was updated in rvcmds
- My commit history is messy. I can clean it up if necessary prior to
merge but wanted to get this out there before more merge conflicts
arise.
- I did not address CI for building 2025 and 2026. That still needs to
be configured.

### Other discoveries
RV builds on macOS still want to find things in /opt/homebrew. This can
easily cause build errors for users that have things like openexr
installed through homebrew. You can see this by greping for
`/opt/homebrew` in the `_build` directory. I tried various cmake vars to
tell cmake to ignore /opt/homebrew but could never seem to get any to
work. Additionally there are things in homebrew that the build relies on
like xz so you don't want to ignore everything. The problem is if
something like openexr4.3.0 is installed with homebrew, cmake for 2024
will attempt to link to that rather than the 3.2 that was compiled for
that CY year. For me I had to uninstall a bunch to get the build to work
properly
```
brew uninstall jpeg-xl libheif imagemagick aom openexr openexr openssl boost ffmpeg imath libtiff
```

The resulting size of the RV.app build is quite large. I believe it can
be significantly trimmed down and will do that in a different PR.

### If possible, provide screenshots.

---------

Signed-off-by: Michael Oliver <mcoliver@gmail.com>