SG-40370: Fix Y-Inverted Color Picker (#849)
### Summarize your change.

Fixes the color inspector from calculating the wrong coordinate when a
RVFileSources has multiple media paths with different orientations

We use multiple media paths to add audio to images sequences from a
proxy mov file.

```
rv.commands.addSource(["/path/to/image.####.exr", "/path/to/proxy_with_audio.mov"])
```

The current logic is to use the `ImageStructureInfo` of last movie in
the list and not what is currently visible. In some cases the second
path might have a different coordinate system then the image sequence.
In our cases the Y coordinate is opposite causing the color picker
sample location to be upside down.

This change reuses the exisiting `getMediaFromContext` method to get the
same media that is displayed.

### Describe what you have tested and on which operating system.
Linux, macOS

Signed-off-by: Mark Reid <markreid@netflixanimation.com>
Signed-off-by: Bernard Laberge <117092886+bernie-laberge@users.noreply.github.com>
Co-authored-by: pbergeron-adsk <patrick.bergeron@autodesk.com>
Co-authored-by: Bernard Laberge <117092886+bernie-laberge@users.noreply.github.com>