1004: Update window title on source media rep change - SG-41311 (#1005)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->
Fixes #1004

### Summarize your change.
Add listener to "source-media-rep-activated"

### Describe the reason for the change.
Window title doesnt update when changing source media reps

### Describe what you have tested and on which operating system.
tested on linux and windows

---------

Signed-off-by: Mirco Tornow <mirco.tornow@outlook.com>