SG-40829:  Better about box (#1000)
### Linked issues
SG-40829

note: needs OTIO version found in python3.cmake file

### Summarize your change.

The current RV/OpenRV dialog box is really plain, it barely shows the
current build hash, but that's basically it. It used to be, in the days
before OpenRV, that the about box contained all the info about the
dependencies and third party components. This was linked to a build
system that was overhauled when going OpenRV, and it never got fixed.

Instead of bringing the old code back, because we have a new build
system anyway, we re-developed the feature from scratch.

During the build phase, a new about_rv.cpp file is created from a python
script executed from within the makefiles, which basically just exports
a html string containing the information of the build variables.

**NOTE: I apologize for the auto-reformatting that happens in the cmake
files when committing. There isn't much I can do about this, it makes
the actual changes very hard to find... :(**

### Describe the reason for the change.

From a QA perspective, this is much better to be able to identify a
build version.

### Describe what you have tested and on which operating system.

macOS, for RV-only.

Even though it should work just fine, this really needs to be tested for
an OpenRV build, and on Windows/Linux too.

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

<img width="543" height="493" alt="image"
src="https://github.com/user-attachments/assets/6a854bec-26c7-4905-8984-ae4413328915"
/>

---------

Signed-off-by: Patrick Bergeron <patrick.bergeron@autodesk.com>
Signed-off-by: pbergeron-adsk <patrick.bergeron@autodesk.com>