SG-40514: Add env var to bypass Clear All confirmation dialog (#1017)
### SG-40514: Add env var to bypass Clear All confirmation dialog

### Linked issues
Fixes NA

### Describe the reason for the change.
The reason for this change is that we are adding Annotation automation
tests and in order to test the "Clear Annotations on All The Frames"
functionality, we needed a way to bypass the confirmation dialog.

### Summarize your change.
This commit adds a new env var which when defined, will skip the
confirmation dialog when executing the function : "Clear Annotations on
All The Frames".

`export RV_SKIP_CONFIRMATIONS=1 `

Note that for efficiency, it is evaluated in the annotate_mode
constructor.

### Describe what you have tested and on which operating system.
Successfully tested on macOS.

### Add a list of changes, and note any that might need special
attention during the review.

### If possible, provide screenshots.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>