SG-41182: Add id to Paint OTIO schema (#999)
### [SG-41182](https://jira.autodesk.com/browse/SG-41182): Add id to
Paint OTIO schema

### Summarize your change.
Add an `id` property in the Paint OTIO schema.

### Describe the reason for the change.
This `id` property will be used to identify which stroke to add the
received points when multiple participants of an OTIO-based Live Review
sessions are annotating simultaneously. This small change is needed to
properly serialize and deserialize PAINT_START events.

Signed-off-by: Éloïse Brosseau <eloise.brosseau@autodesk.com>